Updated examples of Textarea documentation
