+++ docs/src/pages/components/cdk-components/textarea/Textarea.jsx
@@ -5,19 +5,15 @@ import Section from "../../common/Section";
 import Example from "../../common/Example";
 import ComponentHeader from "../../common/ComponentHeader";
 import TextareaPropsTable from "./api.jsx";
+import basic from "./examples/basicUsageTextarea";
 import controlled from "./examples/controlledTextarea";
 import uncontrolled from "./examples/uncontrolledTextarea";
-import placeholder from "./examples/placeholderTextarea";
 import optional from "./examples/optionalTextarea";
-import disabled from "./examples/disabledTextarea";
+import verticalGrow from "./examples/verticalGrowTextarea";
+import rows from "./examples/rowsTextarea";
 import pattern from "./examples/patternTextarea";
 import length from "./examples/lengthTextarea";
-import invalid from "./examples/invalidTextarea";
-import fillParent from "./examples/fillParentTextarea";
 import customErrors from "./examples/customErrorsTextarea";
-import vgManual from "./examples/vgManualTextarea";
-import vgNone from "./examples/vgNoneTextarea";
-import rows from "./examples/rowsTextarea";
 
 function Textarea() {
   return (
@@ -29,24 +25,15 @@ function Textarea() {
       </Section>
       <Section>
         <DxcHeading level={3} text="Examples" margin={{ bottom: "small" }} />
+        <Example title="Basic usage" example={basic}></Example>
         <Example title="Controlled textarea" example={controlled} />
         <Example title="Uncontrolled textarea" example={uncontrolled} />
-        <Example
-          title="Textarea with placeholder"
-          example={placeholder}
-        ></Example>
-        <Example title="Disabled textarea" example={disabled}></Example>
         <Example title="Optional textarea" example={optional}></Example>
         <Example
-          title="Textarea with verticalGrow set to 'manual'"
-          example={vgManual}
-        ></Example>
-        <Example
-          title="Textarea with verticalGrow set to 'none'"
-          example={vgNone}
+          title="Textarea vertical grow"
+          example={verticalGrow}
         ></Example>
         <Example title="Textarea with default rows" example={rows}></Example>
-        <Example title="Invalid textarea" example={invalid}></Example>
         <Example
           title="Textarea with pattern constraint"
           example={pattern}
@@ -58,10 +45,6 @@ function Textarea() {
         <Example
           title="Textarea with custom error messages"
           example={customErrors}
-        ></Example>
-        <Example
-          title="Fill parent size textarea"
-          example={fillParent}
         ></Example>
       </Section>
     </ComponentDoc>
