--- docs/src/pages/components/cdk-components/textarea/examples/disabledTextarea.js
@@ -1,17 +1,0 @@
-import { DxcTextarea } from "@dxc-technology/halstack-react";
-
-const code = `() => {
-  return (
-    <DxcTextarea
-      label="Disabled"
-      disabled
-      margin="medium"
-    />
-  );
-}`;
-
-const scope = {
-  DxcTextarea,
-};
-
-export default { code, scope };
