--- docs/src/pages/components/cdk-components/textarea/examples/fillParentTextarea.js
@@ -1,19 +1,0 @@
-import { DxcTextarea } from "@dxc-technology/halstack-react";
-import { useState } from "react";
-
-const code = `() => {
-  return (
-    <DxcTextarea
-      label="Fill parent"
-      size="fillParent"
-      margin="medium"
-    />
-  );
-}`;
-
-const scope = {
-  DxcTextarea,
-  useState,
-};
-
-export default { code, scope };
