--- docs/src/pages/components/cdk-components/textarea/examples/invalidTextarea.js
@@ -1,19 +1,0 @@
-import { DxcTextarea } from "@dxc-technology/halstack-react";
-import { useState } from "react";
-
-const code = `() => {
-  return (
-    <DxcTextarea
-      label="Invalid"
-      error="Error message."
-      margin="medium"
-    />
-  );
-}`;
-
-const scope = {
-  DxcTextarea,
-  useState,
-};
-
-export default { code, scope };
