--- docs/src/pages/components/cdk-components/textarea/examples/placeholderTextarea.js
@@ -1,18 +1,0 @@
-import { DxcTextarea } from "@dxc-technology/halstack-react";
-import { useState } from "react";
-
-const code = `() => {
-  return (
-    <DxcTextarea
-      placeholder="Placeholder"
-      margin="medium"
-    />
-  );
-}`;
-
-const scope = {
-  DxcTextarea,
-  useState,
-};
-
-export default { code, scope };
