--- docs/src/pages/components/cdk-components/textarea/examples/vgManualTextarea.js
@@ -1,20 +1,0 @@
-import { DxcTextarea } from "@dxc-technology/halstack-react";
-import { useState } from "react";
-
-const code = `() => {
-  return (
-    <DxcTextarea
-      label="Vertical grow"
-      helperText="'manual'"
-      verticalGrow="manual"
-      margin="medium"
-    />
-  );
-}`;
-
-const scope = {
-  DxcTextarea,
-  useState,
-};
-
-export default { code, scope };
